pyproject.toml
[project]
name = "pangolincompliancesuite"
version = "0.1.0"
description = "A Continuous Compliance and Monitoring Framework Im developing a tool"
authors = [
    {name = "manish59",email = "manishkumar.bobbili3@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "djangorestframework (>=3.15.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pangolin-sdk = {path = "/Users/manishkumarbobbili/PycharmProjects/pangolin_sdk/dist/pangolin_sdk-1.1.1-py3-none-any.whl"}

